Correct role annotations for ghost parameters.
GDP has lots of types that are supposed to be
library-author-controlled, but involve phantom type parameters. This
allows users to unsoundly coerce the type parameters. To fix this
problem, we need explicit role annotations on those phantom type
parameters, changing them to nominal (and thus entirely under the
control of the library author / defining module).